+++ core/src/main/java/com/trivago/cluecumber/core/CluecumberCore.java
@@ -45,8 +45,8 @@ public class CluecumberCore {
         cluecumberEngine.setExpandBeforeAfterHooks(builder.expandBeforeAfterHooks);
         cluecumberEngine.setExpandDocStrings(builder.expandDocStrings);
         cluecumberEngine.setExpandStepHooks(builder.expandStepHooks);
-        cluecumberEngine.setShowOnlyLastRuns(builder.showOnlyLastRuns);
-        cluecumberEngine.setExpandPreviousRuns(builder.expandPreviousRuns);
+        cluecumberEngine.setGroupPreviousScenarioRuns(builder.groupPreviousScenarioRuns);
+        cluecumberEngine.setExpandPreviousScenarioRuns(builder.expandPreviousScenarioRuns);
         cluecumberEngine.setFailScenariosOnPendingOrUndefinedSteps(builder.failScenariosOnPendingOrUndefinedSteps);
         cluecumberEngine.setLogLevel(builder.logLevel);
         cluecumberEngine.setStartPage(builder.startPage);
@@ -80,8 +80,8 @@ public class CluecumberCore {
         private boolean expandBeforeAfterHooks;
         private boolean expandDocStrings;
         private boolean expandStepHooks;
-        private boolean showOnlyLastRuns;
-        private boolean expandPreviousRuns;
+        private boolean groupPreviousScenarioRuns;
+        private boolean expandPreviousScenarioRuns;
         private String logLevel;
         private String startPage;
         private boolean failScenariosOnPendingOrUndefinedSteps;
@@ -231,24 +231,24 @@ public class CluecumberCore {
         }
 
         /**
-         * Whether to show the toggle to show elements that are not last run.
+         * Whether to show the scenarios run multiple times should be grouped and the show not last run toggle should be shown.
          *
-         * @param showOnlyLastRuns If true, the toggle to show elements that are not last run will be shown.
+         * @param groupPreviousScenarioRuns If true, the scenarios run multiple times should be grouped and the show not last run toggle should be shown.
          * @return The {@link Builder}.
          */
-        public Builder setShowOnlyLastRuns(final boolean showOnlyLastRuns) {
-            this.showOnlyLastRuns = showOnlyLastRuns;
+        public Builder setGroupPreviousScenarioRuns(final boolean groupPreviousScenarioRuns) {
+            this.groupPreviousScenarioRuns = groupPreviousScenarioRuns;
             return this;
         }
 
         /**
          * Whether to expand elements that are not last run.
          *
-         * @param expandPreviousRuns If true, elements that are not last run will be expanded.
+         * @param expandPreviousScenarioRuns If true, elements that are not last run will be expanded.
          * @return The {@link Builder}.
          */
-        public Builder setExpandPreviousRuns(final boolean expandPreviousRuns) {
-            this.expandPreviousRuns = expandPreviousRuns;
+        public Builder setExpandPreviousScenarioRuns(final boolean expandPreviousScenarioRuns) {
+            this.expandPreviousScenarioRuns = expandPreviousScenarioRuns;
             return this;
         }