+++ engine/src/main/java/com/trivago/cluecumber/engine/rendering/pages/renderering/ScenarioDetailsPageRenderer.java
@@ -76,8 +76,8 @@ public class ScenarioDetailsPageRenderer extends PageWithChartRenderer {
         scenarioDetailsPageCollection.setExpandStepHooks(propertyManager.isExpandStepHooks());
         scenarioDetailsPageCollection.setExpandDocStrings(propertyManager.isExpandDocStrings());
         scenarioDetailsPageCollection.setExpandAttachments(propertyManager.isExpandAttachments());
-        scenarioDetailsPageCollection.setExpandPreviousRuns(propertyManager.isExpandPreviousRuns());
-        scenarioDetailsPageCollection.setShowOnlyLastRuns(propertyManager.isShowOnlyLastRuns());
+        scenarioDetailsPageCollection.setExpandPreviousScenarioRuns(propertyManager.isExpandPreviousScenarioRuns());
+        scenarioDetailsPageCollection.setGroupPreviousScenarioRuns(propertyManager.isGroupPreviousScenarioRuns());
 
         addChartJsonToReportDetails(scenarioDetailsPageCollection);