+++ CHANGELOG.md
@@ -10,8 +10,8 @@ Back to [Readme](README.md).
 ## [3.5.0] - 2023-11-02
 
 ### Added
-* `showOnlyLastRuns` mode for a compact view when there are multiple runs of the same scenario [#245]
-* `expandPreviousRuns` to set default state of expanded or collapsed of the previous runs of the same scenario
+* `groupPreviousScenarioRuns` mode for a compact view when there are multiple runs of the same scenario [#245]
+* `expandPreviousScenarioRuns` to set default state of expanded or collapsed of the previous runs of the same scenario
 
 ## [3.4.0] - 2023-08-10